fix the socket break in the parent process when EM::fork_reactor is called.
The opened sockets prior to calling Kernel.fork are also inherited to the forked child
process (see man 2 fork). EM::fork_reactor causes the socket unbind in the parent process
while the spawn child cleans those inherited sockets up.

In the C EventMachine, the cleanup stage makes two syscalls to the sockets: shutdown()
and close(). These syscalls happens in the child process. However, the shutdown(),
in EventableDescriptor::Close(), to those inherited sockets affects to the parent.

This patch is to eliminate the problem. Added a test case to replay the case.